separeted components and added unfollow clickhandler to each following user
