--- client/src/components/FollowerItem.jsx
@@ -1,63 +1,0 @@
-import React from 'react';
-import { Typography, Grid, Paper } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    margin: 'auto',
-    maxWidth: 500,
-  },
-  image: {
-    width: 128,
-    height: 128,
-  },
-  img: {
-    margin: 'auto',
-    display: 'block',
-    maxWidth: '100%',
-    maxHeight: '100%',
-  },
-}));
-
-function FollowerItem(props) {
-  const classes = useStyles();
-  const { book, handleRemoveClick } = props;
-  console.log(book);
-  return (
-    <div>
-      <Paper className={classes.paper}>
-        <Grid container spacing={2}>
-          <Grid item>
-            <img className={classes.img} alt="complex" src={book.coverURL} />
-          </Grid>
-          <Grid item xs={12} sm container>
-            <Grid item xs container direction="column" spacing={2}>
-              <Grid item xs>
-                <Typography gutterBottom variant="subtitle1">
-                  {book.title}
-                </Typography>
-                <Typography variant="body2" gutterBottom>
-                  {book.author}
-                </Typography>
-                <Typography variant="body2" color="textSecondary">
-                  {book.genre}
-                </Typography>
-              </Grid>
-              <Grid item>
-                <Typography color="primary" variant="body2" style={{ cursor: 'pointer' }} onClick={() => handleRemoveClick(book.isbn, false)}>
-                  Remove from To-Read List
-                </Typography>
-              </Grid>
-            </Grid>
-          </Grid>
-        </Grid>
-      </Paper>
-    </div>
-  );
-}
-
-export default FollowerItem;
